Add `registerClass` command / update to version 1.1.0-alpha.1
New command `registerClass` that will register a single class object into the index.  It can be any object that can be used as a behavior.  Type checking is not performed, but an existence check is.  Re-registering a class with a different long ID is an error unless the second parameter is true (`pUpdateClass`).

Updated the Particles demo stack to use this new function.  The Emittor class was moved to a stack script.  The stack is in the same binary file to make distribution easier, but it could easily be a script only stack.

Switched to Semantic Versioning 2.0.0 standard (semver.org) as 1.1.0-alpha.1
Since functionality has been added that shouldn't break existing code, incrementing the MINOR version