Add object types and single parameter support
Add the idea from Mark's code to allow the object type to be specified for a new object.  Implemented a bit differently.
Added a single parameter for the constructor (`pParams`).  Could implement the multiple parameter code used in `messageObject` but not sure if it is worth the effort and possible performance impact.
Updated `registerClasses` to use `registerClass`
Added `sClassListA` as a cache for the class hierarchy of each class.  Private function `_getClassList` implemented to provide the data.